dtls: make the interface index in the session a u8
It is very unlikely that we have a device with more than 256 Ethernet
interfaces.

+++ global.h
@@ -81,7 +81,7 @@ typedef struct __session_t {
     struct sockaddr_in  sin;
     struct sockaddr_in6 sin6;
   } addr;
-  int ifindex;
+  uint8_t ifindex;
 } __session_t;
 
 #define session_t __session_t
